Refactorizare

+++ src/main/java/com/LoRaWan/LoRaWan/Service/NodeService.java
@@ -2,6 +2,7 @@ package com.LoRaWan.LoRaWan.Service;
 
 import com.LoRaWan.LoRaWan.Date.*;
 import com.LoRaWan.LoRaWan.Dto.*;
+import com.LoRaWan.LoRaWan.Exception.ExceptionNodNotFound;
 import com.LoRaWan.LoRaWan.Repository.*;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -16,6 +17,14 @@ import java.util.*;
 @Service
 public class NodeService {
 
+    @Autowired
+    private HumidityRepository humidityRepository;
+
+    @Autowired
+    private AirQualityRepository airQualityRepository;
+
+    @Autowired
+    private TemperatureRepository temperatureRepository;
 
     @Autowired
     private NodeRepository nodeRepository;
@@ -40,10 +49,6 @@ public class NodeService {
 
     public Node getNodeByDevEui(String devEui) {
         return nodeRepository.getNodeByDevEui(devEui);
-    }
-
-    public List<Node> getNodesByDevEui(String devEui) {
-        return nodeRepository.getNodesByDevEui(devEui);
     }
 
     public void saveNode(NodeDto nodeDto) throws Exception {
@@ -81,78 +86,93 @@ public class NodeService {
         } else {
             throw new Exception("NAME_EXCEPTION");
         }
-
-
     }
 
     public List<NodeDto> getAllNode() {
         List<NodeDto> nodeDtos = new ArrayList<>();
-
         List<Node> list = nodeRepository.findAll();
 
         for (Node node : list) {
-
-            nodeDtos.add(new NodeDto(node.getName(), node.getDevEui(), node.getAppEui(), node.getLocation().getLatitude(),
-                    node.getLocation().getLongitude()));
+            nodeDtos.add(new NodeDto(
+                    node.getName(),
+                    node.getDevEui(),
+                    node.getAppEui(),
+                    node.getLocation().getLatitude(),
+                    node.getLocation().getLongitude())
+            );
         }
-
         return nodeDtos;
     }
 
-    public NodeDto getNode(String name) {
+    public NodeDto getNode(String name) throws ExceptionNodNotFound {
         Node node = nodeRepository.getNodeByName(name);
-        System.out.println(node.getName());
-        return new NodeDto(node.getName(), node.getDevEui(), node.getAppEui(), node.getLocation().getLatitude(), node.getLocation().getLongitude());
+        if (node == null) {
+            throw new ExceptionNodNotFound("Get node not found!!");
+        }
+        return new NodeDto(node.getName(),
+                node.getDevEui(),
+                node.getAppEui(),
+                node.getLocation().getLatitude(),
+                node.getLocation().getLongitude()
+        );
     }
 
-    public DataSenzorDto sendData(String name) {
-        List<Float> temperature = nodeRepository.getTemperatureByNameNode(name);
-        List<Float> humidity = nodeRepository.getHumidityByNameNode(name);
-        List<Float> airQuality = nodeRepository.getAirQualityByNameNode(name);
+    public DataSenzorDto sendData(String name) throws ExceptionNodNotFound {
+        List<Float> temperature = temperatureRepository.getTemperatureByNameNode(name);
+        List<Float> humidity = humidityRepository.getHumidityByNameNode(name);
+        List<Float> airQuality = airQualityRepository.getAirQualityByNameNode(name);
 
+        if (temperature == null || humidity == null || airQuality == null) {
+            throw new ExceptionNodNotFound("Nod not found!!!");
+        }
 
         return new DataSenzorDto(humidity, temperature, airQuality);
     }
 
-
-    public List<NodeByUserDto> getNodeByUser(String email) {
+    public List<NodeByUserDto> getNodeByUser(String email) throws ExceptionNodNotFound {
         List<NodeByUserDto> nodes = new ArrayList<>();
 
         List<Node> nodeList = nodeRepository.getNodeByUser(email);
+        if (nodeList != null) {
+            for (Node node : nodeList) {
+                NodeByUserDto node1 = new NodeByUserDto();
 
-        for (Node node : nodeList) {
-            NodeByUserDto node1 = new NodeByUserDto();
-
-            node1.setAppEui(node.getAppEui());
-            node1.setDevEui(node.getDevEui());
-            node1.setName(node.getName());
-            node1.setBattery(node.getStatus().getBatteryPower().toString());
-            if (node.getStatus().getBatteryPower() > 20) {
-                node1.setActive(true);
-            } else {
-                node1.setActive(false);
+                node1.setAppEui(node.getAppEui());
+                node1.setDevEui(node.getDevEui());
+                node1.setName(node.getName());
+                node1.setBattery(node.getStatus().getBatteryPower().toString());
+                if (node.getStatus().getBatteryPower() > 20) {
+                    node1.setActive(true);
+                } else {
+                    node1.setActive(false);
+                }
+                nodes.add(node1);
             }
-            nodes.add(node1);
+        } else {
+            throw new ExceptionNodNotFound("Nod not found!!!");
         }
 
         return nodes;
     }
 
-    public MyStationDto getMyStation(String email) {
+    public MyStationDto getMyStation(String email) throws ExceptionNodNotFound {
         MyStationDto myStationDto = new MyStationDto();
+        if (nodeRepository.getNodeByUser(email) != null) {
+            myStationDto.setNumberOfStation(nodeRepository.getNodeByUser(email).size());
+            myStationDto.setStationActive(nodeRepository.getNodeByStationActive(email).size());
+            myStationDto.setStationInactive(nodeRepository.getNodeByStationInActive(email).size());
 
-        myStationDto.setNumberOfStation(nodeRepository.getNodeByUser(email).size());
-        myStationDto.setStationActive(nodeRepository.getNodeByStationActive(email).size());
-        myStationDto.setStationInactive(nodeRepository.getNodeByStationInActive(email).size());
-
-        List<Float> humidity = nodeRepository.getHumidityByAllNode(email);
-        myStationDto.setHumidity((int) (sumList(humidity) / humidity.size()));
-        List<Float> temperature = nodeRepository.getTemperatureByAllNode(email);
-        myStationDto.setTemperature((int) (sumList(temperature) / temperature.size()));
-        List<Float> airQuality = nodeRepository.getAirQualityByAllNode(email);
-        myStationDto.setAirQuality((int) (sumList(airQuality) / airQuality.size()));
-        List<Float> batteryPower = nodeRepository.getBatteryPowerNode(email);
-        myStationDto.setBatteryLeve((int) (sumList(batteryPower) / batteryPower.size()));
+            List<Float> humidity = humidityRepository.getHumidityByAllNode(email);
+            myStationDto.setHumidity((int) (sumList(humidity) / humidity.size()));
+            List<Float> temperature = temperatureRepository.getTemperatureByAllNode(email);
+            myStationDto.setTemperature((int) (sumList(temperature) / temperature.size()));
+            List<Float> airQuality = airQualityRepository.getAirQualityByAllNode(email);
+            myStationDto.setAirQuality((int) (sumList(airQuality) / airQuality.size()));
+            List<Float> batteryPower = statusRepository.getBatteryPowerNode(email);
+            myStationDto.setBatteryLeve((int) (sumList(batteryPower) / batteryPower.size()));
+        } else {
+            throw new ExceptionNodNotFound("Nod not found!!!");
+        }
         return myStationDto;
     }
 
@@ -166,44 +186,47 @@ public class NodeService {
     }
 
     @Transactional(propagation = Propagation.REQUIRES_NEW)
-    public List<MessagePayloadDto> getNodeByName(String name, int limSup, int limInf) {
-
-        List<AcquisitionDate> messages = messageRepository.getPayloadByName(name, limSup, limInf);
+    public List<MessagePayloadDto> getNodeByName(String name, int limSup, int limInf) throws ExceptionNodNotFound {
         List<MessagePayloadDto> messagePayloadDtos = new ArrayList<>();
-        for (AcquisitionDate date : messages) {
-            MessagePayloadDto messagePayloadDto = new MessagePayloadDto();
-            messagePayloadDto.setAirQuality(date.getAirQ());
-            messagePayloadDto.setHumidity(date.getHum());
-            messagePayloadDto.setTemperature(date.getTemp());
-            messagePayloadDto.setDate(date.getDate());
-            messagePayloadDtos.add(messagePayloadDto);
+        if (nodeRepository.getNodeByName(name) != null) {
+            List<AcquisitionDate> messages = messageRepository.getPayloadByName(name, limSup, limInf);
+
+            for (AcquisitionDate date : messages) {
+                MessagePayloadDto messagePayloadDto = new MessagePayloadDto();
+                messagePayloadDto.setAirQuality(date.getAirQ());
+                messagePayloadDto.setHumidity(date.getHum());
+                messagePayloadDto.setTemperature(date.getTemp());
+                messagePayloadDto.setDate(date.getDate());
+                messagePayloadDtos.add(messagePayloadDto);
+            }
+        } else {
+            throw new ExceptionNodNotFound("Nod not found!!!");
         }
-
-
         return messagePayloadDtos;
     }
 
-    public String deleteNode(String name) {
-
+    public void deleteNode(String name) throws ExceptionNodNotFound {
         Node node = nodeRepository.getNodeByName(name);
         nodeRepository.foreignKey(0);
-
-        List<Message> messages = node.getMessages();
-
-
-        for (Message message : messages) {
-            payloadRepository.delete(message.getPayload());
-            gatewayRepository.delete(message.getGateway());
-
-            messageRepository.delete(message);
+        if (node != null) {
+            try {
+                List<Message> messages = node.getMessages();
+                for (Message message : messages) {
+                    payloadRepository.delete(message.getPayload());
+                    gatewayRepository.delete(message.getGateway());
+                    messageRepository.delete(message);
+                }
+            } catch (NullPointerException exception) {
+                System.out.println("Nodul nu are mesaje");
+            }
+            nodeRepository.delete(node);
+        } else {
+            throw new ExceptionNodNotFound("Nod not found!!!");
         }
-        nodeRepository.delete(node);
-
-        return "da";
     }
 
 
-    public void addNewLocation(NewLocationDto newLocationDto) throws Exception {
+    public void addNewLocation(NewLocationDto newLocationDto) throws ExceptionNodNotFound {
 
         Node node = nodeRepository.getNodeByName(newLocationDto.getName());
 
@@ -217,18 +240,17 @@ public class NodeService {
 
             newLocationRepository.save(newLocation);
         } else {
-            throw new Exception("Nu exista nod!!!");
+            throw new ExceptionNodNotFound("Nod not found!!!");
         }
 
     }
 
-    public List<NewLocationDto> selectAllNewLocation(String email) {
+    public List<NewLocationDto> selectAllNewLocation(String email) throws ExceptionNodNotFound {
 
         User user = userRepository.getUserByEmail(email);
         List<NewLocationDto> list = null;
         if (user != null) {
             list = new ArrayList<>();
-
             for (Node node : user.getNodeList()) {
                 for (NewLocation location : node.getNewLocation()) {
                     NewLocationDto newLocationDto = new NewLocationDto();
@@ -238,132 +260,40 @@ public class NodeService {
                     newLocationDto.setLat(location.getLatitude());
                     list.add(newLocationDto);
                 }
-
             }
+        } else {
+            throw new ExceptionNodNotFound("Nod not found!!!");
         }
         return list;
     }
 
 
-    public DateValue getDateLast7Day(String name) {
-        Date date = new Date();
-        Calendar c = Calendar.getInstance();
-        c.setTime(date);
-        c.add(Calendar.DATE, -17);
-
-        Date dateC = c.getTime();
-
-        List<GraphRequest> humidity = nodeRepository.getHumidityForDate(name, dateC);
-        List<GraphRequest> temperature = nodeRepository.getTemperatureForDate(name, dateC);
-        List<GraphRequest> airQuality = nodeRepository.getAirQualityForDate(name, dateC);
-        DateValue dateValue = new DateValue();
-        List<Float> humidityDate = new ArrayList<>();
-        List<Float> temperatureDate = new ArrayList<>();
-        List<Float> airQualityDate = new ArrayList<>();
-        List<Date> integers = new ArrayList<>();
-        for (int i = 0; i < humidity.size(); i++) {
-            humidityDate.add(humidity.get(i).getVal());
-            temperatureDate.add(temperature.get(i).getVal());
-            airQualityDate.add(airQuality.get(i).getVal());
-            integers.add(humidity.get(i).getDate());
-        }
-        dateValue.setHumidity(humidityDate);
-        dateValue.setTemperature(temperatureDate);
-        dateValue.setAirQuality(airQualityDate);
-        dateValue.setDates(integers);
-
-        return dateValue;
-    }
-
-    public DateValue getDateLastMount(String name) {
-        Date date = new Date();
-        Calendar c = Calendar.getInstance();
-        c.setTime(date);
-        c.add(Calendar.MONTH, -1);
-
-        Date dateC = c.getTime();
-        System.out.println(dateC.toString());
-        System.out.println(name);
-        List<GraphRequest> humidity = nodeRepository.getHumidityForDate(name, dateC);
-        List<GraphRequest> temperature = nodeRepository.getTemperatureForDate(name, dateC);
-        List<GraphRequest> airQuality = nodeRepository.getAirQualityForDate(name, dateC);
-        DateValue dateValue = new DateValue();
-
-        List<Float> humidityDate = new ArrayList<>();
-        List<Float> temperatureDate = new ArrayList<>();
-        List<Float> airQualityDate = new ArrayList<>();
-        List<Date> integers = new ArrayList<>();
-        for (int i = 0; i < humidity.size(); i++) {
-            humidityDate.add(humidity.get(i).getVal());
-            temperatureDate.add(temperature.get(i).getVal());
-            airQualityDate.add(airQuality.get(i).getVal());
-            integers.add(humidity.get(i).getDate());
-        }
-        dateValue.setHumidity(humidityDate);
-        dateValue.setTemperature(temperatureDate);
-        dateValue.setAirQuality(airQualityDate);
-        dateValue.setDates(integers);
-
-        return dateValue;
-    }
-
-    public DateValue getDateLastYear(String name) {
-        Date date = new Date();
-        Calendar c = Calendar.getInstance();
-        c.setTime(date);
-        c.add(Calendar.YEAR, -1);
-
-
-        Date dateC = c.getTime();
-        System.out.println(dateC.toString());
-
-        List<GraphRequest> humidity = nodeRepository.getHumidityForYear(name, dateC);
-        List<GraphRequest> temperature = nodeRepository.getTemperatureForYear(name, dateC);
-        List<GraphRequest> airQuality = nodeRepository.getAirQualityForYear(name, dateC);
-        DateValue dateValue = new DateValue();
-
-        List<Float> humidityDate = new ArrayList<>();
-        List<Float> temperatureDate = new ArrayList<>();
-        List<Float> airQualityDate = new ArrayList<>();
-        List<Integer> integers = new ArrayList<>();
-        for (int i = 0; i < humidity.size(); i++) {
-            humidityDate.add(humidity.get(i).getVal());
-            temperatureDate.add(temperature.get(i).getVal());
-            airQualityDate.add(airQuality.get(i).getVal());
-            integers.add(humidity.get(i).getMonth());
-        }
-        dateValue.setHumidity(humidityDate);
-        dateValue.setTemperature(temperatureDate);
-        dateValue.setAirQuality(airQualityDate);
-        dateValue.setMonth(integers);
-
-        return dateValue;
-    }
-
-
-    public List<String> getNodeNameListByUser(String email) {
+    public List<String> getNodeNameListByUser(String email) throws ExceptionNodNotFound {
         List<Node> nodeList = nodeRepository.getNodeByUser(email);
         List<String> nodeNameList = new ArrayList<>();
+        if (nodeList != null) {
 
-        for (Node node : nodeList) {
-            nodeNameList.add(node.getName());
+            for (Node node : nodeList) {
+                nodeNameList.add(node.getName());
+            }
+        } else {
+            throw new ExceptionNodNotFound("Nod not found!!!");
         }
-
         return nodeNameList;
     }
 
-    public List<String> last10Message(String name) {
+    public List<String> last10Message(String name) throws ExceptionNodNotFound {
         List<Message> messageList = messageRepository.getLast10Message(name);
-
         List<String> messageStringList = new ArrayList<>();
-        for (Message message : messageList) {
-
-            message.getGateway().getRssi();
-            String mes = message.getPayload().getRecvPayload().toGMTString().toString() + " rssi:" + message.getGateway().getRssi();
-            messageStringList.add(mes);
+        if (messageList != null) {
+            for (Message message : messageList) {
+                message.getGateway().getRssi();
+                String mes = message.getPayload().getRecvPayload().toGMTString().toString() + " rssi:" + message.getGateway().getRssi();
+                messageStringList.add(mes);
+            }
+        } else {
+            throw new ExceptionNodNotFound("Nod not found!!!");
         }
         return messageStringList;
     }
-
-
 }
